refactor: move enums and dataclasses to models.py
